consolidated Block class
