+++ blocks/aem-columns/aem-columns.js
@@ -1,3 +1,3 @@
-import { Block } from '../../scripts/aem-blocks.js';
+import { Block } from '../../scripts/aem.js';
 
 export default class Columns extends Block {}
