+++ blocks/aem-footer/aem-footer.js
@@ -1,3 +1,3 @@
-import { Block } from '../../scripts/aem-blocks.js';
+import { Block } from '../../scripts/aem.js';
 
 export default class Footer extends Block {}
