+++ blocks/aem-section-metadata/aem-section-metadata.js
@@ -1,8 +1,12 @@
-import { MetaBlock } from '../../scripts/aem-blocks.js';
+import { Block } from '../../scripts/aem.js';
 
-export default class SectionMetadata extends MetaBlock {
+export default class SectionMetadata extends Block {
+  constructor() {
+    super({ mapValues: true });
+  }
+
   connectedCallback() {
-    [...this.data].forEach(([key, value]) => {
+    [...this.values].forEach(([key, value]) => {
       if (key === 'style') {
         this.parentElement.classList.add(value);
       } else {
