+++ head.html
@@ -2,7 +2,6 @@
 
 <script type="module">
     import initialize from '/scripts/aem.js'
-    import '/scripts/aem-blocks.js'
     initialize();
 </script>
